fix(ci): split loan-officer types and clear stale mypy ignore
